Moved possibly reusable components to common
