+++ src/common/components/terminology-components/information-domains-selector.tsx
@@ -3,8 +3,11 @@ import { useState } from 'react';
 import { MultiSelectData } from 'suomifi-ui-components';
 import { useBreakpoints } from '@app/common/components/media-query/media-query-context';
 import { useGetGroupsQuery } from '@app/common/components/terminology-search/terminology-search.slice';
-import { BlankFieldset, MultiselectSmBot } from './new-terminology.styles';
-import { UpdateTerminology } from './update-terminology.interface';
+import {
+  BlankFieldset,
+  MultiselectSmBot,
+} from './terminology-components.styles';
+import { UpdateTerminology } from '@app/modules/new-terminology/update-terminology.interface';
 
 interface InformationDomainsSelectorProps {
   update: ({ key, data }: UpdateTerminology) => void;
