+++ src/common/components/terminology-components/language-selector.tsx
@@ -7,8 +7,8 @@ import {
   BlankFieldset,
   BlankLegend,
   MultiselectSmBot,
-} from './new-terminology.styles';
-import { UpdateTerminology } from './update-terminology.interface';
+} from './terminology-components.styles';
+import { UpdateTerminology } from '@app/modules/new-terminology/update-terminology.interface';
 
 export interface TerminologyName {
   lang: string;
